+++ scripts/lichessTools-init.js
@@ -62,6 +62,7 @@ lichessTools.loadTool(LiChessTools.Tools.TranspositionArrows);
 lichessTools.loadTool(LiChessTools.Tools.TvOptions);
 lichessTools.loadTool(LiChessTools.Tools.ShowPawnStructure);
 lichessTools.loadTool(LiChessTools.Tools.ExplorerSnaps);
+lichessTools.loadTool(LiChessTools.Tools.MiniGameSize);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);
 lichessTools.loadTool(LiChessTools.Tools.TrapValueCommand);